+++ background_scripts/main.js
@@ -101,6 +101,43 @@ Links = {
   }
 };
 
+Files = {
+  sendRequest: function(path, callback) {
+    var xhr;
+    xhr = new XMLHttpRequest();
+    xhr.open("GET", "file://" + path);
+    xhr.onreadystatechange = function() {
+      if (xhr.readyState === 4) {
+        callback(xhr.responseText);
+      }
+    };
+    xhr.send();
+  },
+
+  parseHTML: function(data) {
+    var matches = data.match(/addRow\("[^)]+"\)/g);
+    var results = [];
+    if (matches) {
+      for (var i = 0, l = matches.length; i < l; ++i) {
+        var m = JSON.parse(matches[i].replace(/[^(]+\(/, "[").slice(0, -1) + "]");
+        results.push([m[0], m[2]]);
+      }
+    }
+    return results;
+  },
+
+  getPath: function(path, callback) {
+    path = path.replace(/[^\/]*$/, "");
+    if (!path) {
+      return;
+    }
+    this.sendRequest(path, function(data) {
+      data = Files.parseHTML(data);
+      callback(data);
+    });
+  }
+};
+
 Bookmarks = {
   getMarks: function(callback) {
     chrome.bookmarks.getTree(function(tree) {
@@ -187,6 +224,10 @@ chrome.commands.onCommand.addListener(function(command) {
       chrome.tabs.sendMessage(tab[0].id, {action: "nextCompletionResult"}, function() {
         chrome.windows.create({url: "chrome://newtab"});
       });
+    });
+  } else if (command === "closeTab") {
+    chrome.tabs.query({active: true, currentWindow: true}, function(tab) {
+      chrome.tabs.remove(tab[0].id);
     });
   }
 });