+++ content_scripts/utils.js
@@ -1,3 +1,34 @@
+function isValidB64(a) {
+  try {
+    window.atob(a);
+  } catch(e) {
+    return false;
+  }
+  return true;
+}
+
+function reverseImagePost(url) {
+  return '<html><head><title>cVim reverse image search</title></head><body><form id="f" method="POST" action="https://www.google.com/searchbyimage/upload" enctype="multipart/form-data"><input type="hidden" name="image_content" value="' + url.substring(url.indexOf(",") + 1).replace(/\+/g, "-").replace(/\//g, "_").replace(/\./g, "=") + '"><input type="hidden" name="filename" value=""><input type="hidden" name="image_url" value=""><input type="hidden" name="sbisrc" value=""></form><script>document.getElementById("f").submit();\x3c/script></body></html>';
+}
+
+// Based off of the "Search by Image" Chrome Extension by Google
+function googleReverseImage(url, source, c, d) {
+  if (void 0 !== url && url.indexOf("data:") === 0) {
+    if (url.search(/data:image\/(bmp|gif|jpe?g|png|webp|tiff|x-ico)/i) === 0) {
+      var commaIndex = url.indexOf(",");
+      if (commaIndex !== -1 && isValidB64(url.substring(commaIndex + 1))) {
+        return "data:text/html;charset=utf-8;base64," + window.btoa(reverseImagePost(url, source));
+      }
+    }
+  } else {
+    if (url.indexOf("file://") === 0 || url.indexOf("chrome") === 0) {
+      chrome.runtime.sendMessage({action: "urlToBase64", url: url});
+      return;
+    }
+    return "https://www.google.com/searchbyimage?image_url=" + url;
+  }
+}
+
 HTMLElement.prototype.isInput = function() {
   return (
       (this.localName === "textarea" || this.localName === "input" || this.getAttribute("contenteditable") === "true") && !this.disabled &&
@@ -10,7 +41,7 @@ function getVisibleBoundingRect(node) {
   if (computedStyle.opacity === "0" || computedStyle.visibility !== "visible" || computedStyle.display === "none") {
     return false;
   }
-  var boundingRect = node.getBoundingClientRect();
+  var boundingRect = node.getClientRects()[0] || node.getBoundingClientRect();
   if (boundingRect.top > window.innerHeight || boundingRect.left > window.innerWidth) {
     return false;
   }
@@ -21,7 +52,7 @@ function getVisibleBoundingRect(node) {
     var children = node.children;
     var visibleChildNode = false;
     for (var i = 0, l = children.length; i < l; ++i) {
-      boundingRect = children[i].getBoundingClientRect();
+      boundingRect = children[i].getClientRects()[0] || children[i].getBoundingClientRect();
       if (boundingRect.width || boundingRect.height) {
         visibleChildNode = true;
         break;
@@ -55,15 +86,6 @@ function isClickable(node, type) {
         ((t = node.getAttribute("role")) && (t === "button" || t === "checkbox" || t === "menu"))) {
     return true;
   }
-  // if (name === "div" && node.className === "fc-panel") {
-  //   return true;
-  // }
-  // if (getComputedStyle(node).cursor === "pointer") {
-  //   if (!node.offsetParent || (node.offsetParent.style.cursor !== "pointer" && node.offsetParent.localName !== "a" &&
-  //         !/aria|jsaction|role|onclick/.test(Array.prototype.slice.call(node.offsetParent.attributes).map(function(e){return e.name;}).join(",")))) {
-  //     return true;
-  //   }
-  // }
 }
 
 HTMLCollection.prototype.toArray = function() {